backend/requirements.txt
dashscope==1.25.1
fastapi==0.121.2
slowapi==0.1.9
cachetools==5.5.0
langchain==1.2.10
langchain-community==0.4.1
langchain-litellm==0.6.4
langchain-qdrant==1.1.0
langchain-tavily==0.2.17
langgraph-checkpoint-postgres==3.0.4
langgraph-checkpoint-sqlite==3.0.3
langgraph==1.0.9
langsmith==0.4.59
python-dotenv==1.2.1
asyncpg==0.31.0
aiosqlite==0.22.1
psycopg-binary==3.3.2
sqlite-vec-0.1.9
uvicorn==0.38.0
cryptography==46.0.7
